Create method for createBooking, searchAvailableRoom. The logic for booking a team has been written.

+++ ConferenceRoomsWebAPI/DTO/Incoming/CompanyServiceRequest.cs
@@ -3,6 +3,6 @@
     public class CompanyServiceRequest
     {
         public string ServiceName { get; set; } = null!;
-        public double PriceService { get; set; }
+        public decimal PriceService { get; set; }
     }
 }
